requirements.txt
faiss-cpu
numpy
requests
python-dotenv